refactor: rename cli.py to _cli.py
