+++ src/taipy/cli/__init__.py
@@ -9,4 +9,4 @@
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-from .cli import _CLI
+from ._cli import _CLI
